Moved all author and post related code into their own folders. Using graphql-tool/load-files to dynamically load schemas and resolvers from all resources types
