+++ src/authors/authors.resolvers.ts
@@ -1,15 +1,11 @@
 import { IResolvers } from '@graphql-tools/utils';
-import { Author, IAuthor } from '../../models/author';
-import { validateAuthorInput } from '../../utils/validation';
-import { IPost, Post } from '../../models/post';
+import { Author, IAuthor } from './authors.model';
+import { validateAuthorInput } from './authors.validation';
 
 const resolvers: IResolvers = {
   Query: {
     author: async (_: any, { id }: any): Promise<IAuthor | null> => {
       return await Author.findById(id);
-    },
-    post: async (_: any, { id }: any): Promise<IPost | null> => {
-      return await Post.findById(id);
     },
   },
   Mutation: {
@@ -33,17 +29,6 @@ const resolvers: IResolvers = {
     },
     deleteAuthor: async (_: any, { id }: any): Promise<IAuthor | null> => {
       return await Author.findByIdAndDelete(id);
-    },
-    createPost: async (_: any, { input }: any): Promise<IPost> => {
-      const post = new Post(input);
-      await post.save();
-      return post;
-    },
-    updatePost: async (_: any, { id, input }: any): Promise<IPost | null> => {
-      return await Post.findByIdAndUpdate(id, input, { new: true });
-    },
-    deletePost: async (_: any, { id }: any): Promise<IPost | null> => {
-      return await Post.findByIdAndDelete(id);
     },
   },
   Author: {
