+++ src/authors/authors.schema.graphql
@@ -19,26 +19,3 @@ type Mutation {
   updateAuthor(id: ID!, input: AuthorInput!): Author
   deleteAuthor(id: ID!): Author
 }
-
-type Post {
-  id: ID!
-  title: String!
-  content: String!
-  author: Author
-}
-
-input PostInput {
-  title: String!
-  content: String!
-  authorId: ID
-}
-
-type Query {
-  post(id: ID!): Post
-}
-
-type Mutation {
-  createPost(input: PostInput!): Post
-  updatePost(id: ID!, input: PostInput!): Post
-  deletePost(id: ID!): Post
-}
