+++ src/posts/posts.schema.graphql
@@ -1,0 +1,22 @@
+type Post {
+  id: ID!
+  title: String!
+  content: String!
+  author: Author
+}
+
+input PostInput {
+  title: String!
+  content: String!
+  authorId: ID
+}
+
+type Query {
+  post(id: ID!): Post
+}
+
+type Mutation {
+  createPost(input: PostInput!): Post
+  updatePost(id: ID!, input: PostInput!): Post
+  deletePost(id: ID!): Post
+}
