+++ src/posts/posts.validation.ts
@@ -1,0 +1,16 @@
+import { IPost } from './posts.model';
+import validator from 'validator';
+
+export const validatePostInput = (postInput: Partial<IPost>): string[] => {
+  const errors: string[] = [];
+
+  if (!postInput.title || postInput.title.trim().length === 0) {
+    errors.push('Title is required');
+  }
+
+  if (!postInput.content || postInput.content.trim().length === 0) {
+    errors.push('Content is required');
+  }
+
+  return errors;
+};
